ofxFaceTracker Bug fixes
-Fixed crash on iOS isolated to the setup routine of
ofxFaceTrackerThreaded not completing before the rest of the app
query’s the threaded tracker which may lead to crash if variables were
NULL or 0 in size. (This may also happen on older / slower machines or
with a slow HDD). Fix proposed just returns the default datatype as was
setup (stops null vector crashes etc).
-Fixed uninitialised variable ‘failedMiddle’. Causing a false positive
reading for the tracker.getFound() in the first loop in LLVM / iOS.
Making sure it is a true (so YES to fail) at the start fixes this.