ofxFaceTrackerThreaded fixes
Fixes issues #74

 - Fixed crash on startup for ofxFaceTrackerThreaded for iOS/Slow
devices by returning default datatypes instead of clashing against
NULL's or empty vectors.
Problem was the ofxFaceTrackerThreaded::setup was not completing before
the rest of the app started using it. So if the oF app running in the
other thread asks the ofxFaceTracker for information via a function
which may lead to crash if variables were NULL or 0 in size.

 - Fixed uninitialised variable ‘failedMiddle’.
Problem: Causing a false positive reading for the tracker.getFound() in
the first loop in LLVM / iOS.
Solution: Making sure it is a true (so YES to fail) at the start fixes
this.